# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IResponse_any_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{},
							{
								"items": {},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"GeneralFileResponseDto": {
				"properties": {
					"fileGuid": {
						"type": "string"
					},
					"fileName": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"contentType": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fileGuid",
					"fileName",
					"fileUrl",
					"contentType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GetPropertyListDto": {
				"properties": {
					"floorId": {
						"type": "string"
					},
					"units": {
						"items": {
							"properties": {
								"assignedTo": {
									"type": "string",
									"nullable": true
								},
								"isAssigned": {
									"type": "boolean"
								},
								"unitId": {
									"type": "string"
								}
							},
							"required": [
								"assignedTo",
								"isAssigned"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"floorId",
					"units"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetPropertyListDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetPropertyListDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetPropertyListDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"GeneralFileDto": {
				"properties": {
					"fileName": {
						"type": "string"
					},
					"fileData": {
						"type": "string"
					},
					"contentType": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"fileName",
					"fileData",
					"contentType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateParcelDto": {
				"properties": {
					"parcelImage": {
						"$ref": "#/components/schemas/GeneralFileDto"
					},
					"floor": {
						"type": "string"
					},
					"unit": {
						"type": "string"
					}
				},
				"required": [
					"parcelImage",
					"floor",
					"unit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetParcelDto": {
				"properties": {
					"parcelId": {
						"type": "number",
						"format": "double"
					},
					"parcelGuid": {
						"type": "string"
					},
					"parcelImage": {
						"$ref": "#/components/schemas/GeneralFileResponseDto"
					},
					"floor": {
						"type": "string"
					},
					"unit": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"createdDateTime": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"updatedDateTime": {
						"type": "string"
					}
				},
				"required": [
					"parcelId",
					"parcelGuid",
					"parcelImage",
					"floor",
					"unit",
					"createdBy",
					"createdDateTime",
					"updatedBy",
					"updatedDateTime"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedResponse_GetParcelDto_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"data": {
						"properties": {
							"count": {
								"type": "number",
								"format": "double"
							},
							"list": {
								"items": {
									"$ref": "#/components/schemas/GetParcelDto"
								},
								"type": "array",
								"nullable": true
							}
						},
						"required": [
							"count",
							"list"
						],
						"type": "object"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IResponse_GetParcelDto-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"items": {
									"$ref": "#/components/schemas/GetParcelDto"
								},
								"type": "array"
							},
							{
								"items": {
									"items": {
										"$ref": "#/components/schemas/GetParcelDto"
									},
									"type": "array"
								},
								"type": "array"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Secure Guard Pro",
		"version": "1.0.0",
		"description": "Backend API",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/ref-data/properties": {
			"get": {
				"operationId": "getPropertyList",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetPropertyListDto-Array_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetPropertyListDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"RefData"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "checkOccupied",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/parcels/staff": {
			"post": {
				"operationId": "createParcel",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_any_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Parcel"
				],
				"security": [
					{
						"jwt": [
							"STF"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateParcelDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "getParcelByStaff",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetParcelDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetParcelDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Parcel"
				],
				"security": [
					{
						"jwt": [
							"STF"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/parcels": {
			"get": {
				"operationId": "getParcelByResident",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedResponse_GetParcelDto_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse_GetParcelDto-Array_"
								},
								"examples": {
									"Example 1": {}
								}
							}
						}
					}
				},
				"tags": [
					"Parcel"
				],
				"security": [
					{
						"jwt": [
							"RES",
							"SUB"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}